refactor(Explorers): re-organized files for only one explorer
