+++ test/application/explorers/bitcoin/bitpay.test.js
@@ -1,5 +1,5 @@
-import * as mockBitpayResponse from '../../mocks/mockBitpayResponse';
-import { parseTransactionDataFromBitpayResponse } from '../../../../../src/explorers/bitcoin/apis/bitpay';
+import * as mockBitpayResponse from '../mocks/mockBitpayResponse.json';
+import { parseTransactionDataFromBitpayResponse } from '../../../../src/explorers/bitcoin/bitpay';
 
 function getMockBitpayResponse () {
   return JSON.parse(JSON.stringify(mockBitpayResponse));
